Move shared/utils/compare.ts to renderer.
