+++ src/renderer/util/compare.ts
